// contracts/generated/typescript/models/DatasetView.ts
export interface DatasetView {
    /**
     *
     * @type {string}
     * @memberof DatasetView
     */
    id: string;
    /**
     *
     * @type {string}
     * @memberof DatasetView
     */
    datasetId: string;
    /**
     *
     * @type {DatasetViewViewTypeEnum}
     * @memberof DatasetView
     */
    viewType: DatasetViewViewTypeEnum;
    /**
     *
     * @type {string}
     * @memberof DatasetView
     */
    name: string;
    /**
     *
     * @type {DatasetViewConfig}
     * @memberof DatasetView
     */
    config: DatasetViewConfig;
    /**
     *
     * @type {string}
     * @memberof DatasetView
     */
    createdBy: string;
    /**
     *
     * @type {Date}
     * @memberof DatasetView
     */
    createdAt: Date;
}


/**
 * @export
 */
export const DatasetViewViewTypeEnum = {
    Table: 'table',
    Cards: 'cards',
    Chart: 'chart'
} as const;
export type DatasetViewViewTypeEnum = typeof DatasetViewViewTypeEnum[keyof typeof DatasetViewViewTypeEnum];